V2! Start of work on SOFTSETTINGS
this is a V2, this will not work with the last EEPROM version, you will need to update the EEPROM SOFTSETTINGS if you update.

This version makes a big step towords saving everythin in EEPROM, so that 1 firmware version can be used in the future.

TODO: Rotation of keys doesn't seem to fully work yet, will need to do some work on it, firmware has been checked to work on a rotation 1 (normal, no rotation) 4 button PCB